requirements.txt
opencv-python-headless
numpy
Flask
flask-cors
yt-dlp
pycryptodome